# docs/swagger.yaml
basePath: /
definitions:
  httpapi.authSessionResponse:
    properties:
      account_id:
        type: string
      token:
        type: string
    type: object
  httpapi.createFamilyInviteLinkResponse:
    properties:
      expires_at:
        type: string
      invite_url:
        type: string
      token:
        type: string
    type: object
...
  httpapi.errorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  httpapi.joinFamilyByInviteLinkRequest:
    properties:
      member_name:
        type: string
      token:
        type: string
    type: object
  httpapi.joinFamilyByInviteLinkResponse:
    properties:
      family_id:
        type: string
      member_id:
        type: string
    type: object
...
  httpapi.testLoginRequest:
    properties:
      identifier:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Baby sleep tracking API for families.
  title: Keklik API
  version: "1.0"
paths:
  /auth/google/callback:
    get:
      parameters:
      - description: Authorization code returned by Google
        in: query
        name: code
        required: true
        type: string
      - description: State value for CSRF verification
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.authSessionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      summary: Google OAuth callback
      tags:
      - auth
...
  /auth/test/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Test login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.testLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.authSessionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      summary: Test login (dev/staging only)
      tags:
      - auth
...
  /families/invite-links:
    post:
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpapi.createFamilyInviteLinkResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Create family invite link
      tags:
      - families
  /families/join-by-invite-link:
    post:
      consumes:
      - application/json
      parameters:
      - description: Invite token and member name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.joinFamilyByInviteLinkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpapi.joinFamilyByInviteLinkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Join family by invite link
      tags:
      - families
...
securityDefinitions:
  BearerAuth:
    description: Session token obtained from the auth endpoints.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"